move `SelectColumns` from helper.rs to traits.rs
