+++ src/query/helper.rs
@@ -677,38 +677,7 @@ pub trait QueryFilter: Sized {
     }
 }
 
-/// Select specific column for partial model queries
-pub trait SelectColumns {
-    /// Add a select column
-    ///
-    /// For more detail, please visit [QuerySelect::column]
-    fn select_column<C: ColumnTrait>(self, col: C) -> Self;
 
-    /// Add a select column with alias
-    ///
-    /// For more detail, please visit [QuerySelect::column_as]
-    fn select_column_as<C, I>(self, col: C, alias: I) -> Self
-    where
-        C: IntoSimpleExpr,
-        I: IntoIdentity;
-}
-
-impl<S> SelectColumns for S
-where
-    S: QuerySelect,
-{
-    fn select_column<C: ColumnTrait>(self, col: C) -> Self {
-        QuerySelect::column(self, col)
-    }
-
-    fn select_column_as<C, I>(self, col: C, alias: I) -> Self
-    where
-        C: IntoSimpleExpr,
-        I: IntoIdentity,
-    {
-        QuerySelect::column_as(self, col, alias)
-    }
-}
 
 pub(crate) fn join_condition(mut rel: RelationDef) -> Condition {
     // Use table alias (if any) to construct the join condition
